+++ torchinductor/scheduler.py
@@ -9,6 +9,7 @@ from typing import List
 
 import numpy as np
 import torch
+import os
 
 from . import config
 from . import dependencies
@@ -18,6 +19,10 @@ from .dependencies import MemoryDep
 from .dependencies import StarDep
 from .sizevars import SimplifyIndexing
 from .virtualized import V
+from torch.fx.passes.tools_common import legalize_graph
+from functorch._src.partitioners import draw_graph
+from torch.fx.graph_module import GraphModule
+from torch.fx.passes.shape_prop import TensorMetadata
 
 template_kernels = [ir.Convolution]
 
@@ -409,6 +414,103 @@ class NodeUser:
         return self.node.get_name()
 
 
+def get_fake_func(name):
+    def func1(*args):
+        return 0
+    func1.__name__ = name
+    return func1
+
+
+def create_fx_graph(nodes, fname, print_graph = False):
+    """
+    Draw a graph in fname.svg.
+
+    nodes is a list of SchedulerNode objects.
+    """
+    func_dict = {}
+    name_to_fx_node = {}
+    graph = torch.fx.Graph()
+    first_node = None
+
+    # create call_function node for each Buffer and Kernel
+    for snode in nodes:
+        node = snode.node
+        name = node.get_name()      
+        node_type = str(type(node)).split(".")[-1].replace("'>","")
+
+        if node_type in func_dict:
+            fake_f = func_dict[node_type]
+        else:
+            fake_f = get_fake_func(node_type)
+            func_dict[node_type] = fake_f
+        fx_node = graph.call_function(fake_f, args=(), kwargs=None)
+        fx_node.name = name
+
+        # gather meta data
+        dtype = None
+        if isinstance(node, ir.ComputedBuffer):
+            dtype = node.data.dtype
+
+        # try:
+        stride = node.get_stride()
+        layout = type(node.layout)
+
+        if isinstance(snode, NopKernelSchedulerNode):
+            group = "nop"
+        elif isinstance(snode, ExternKernelSchedulerNode):
+            if should_use_template(node):
+                group = snode.group[1]
+            else:
+                group = "extern"
+        else:  # SchedulerNode
+            group = snode.group[1]
+        # except:
+        #     group = torch.Size([0])
+            
+        metadata = TensorMetadata(group, dtype, False, stride, layout, None, None)
+        fx_node.meta["tensor_meta"] = metadata
+
+        name_to_fx_node[name] = fx_node
+        if first_node is None:
+            first_node = fx_node
+
+    # create edges between nodes
+    for snode in nodes:
+        node = snode.node
+        name = node.get_name()      
+        deps = node.get_reads()
+        fx_node = name_to_fx_node[name]
+        
+        new_args = []
+        for dep in deps:
+            if dep.name in name_to_fx_node:
+                dep_node = name_to_fx_node[dep.name]
+            else:
+                with graph.inserting_before(first_node):
+                    dep_node = graph.placeholder(dep.name)  # assume it's a placeholder if not a computebox
+                    name_to_fx_node[dep.name] = dep_node
+            new_args.append(dep_node)
+
+        fx_node.args = tuple(new_args)
+    
+    outputs = []
+    for _,v in name_to_fx_node.items():
+        if len(v.users) == 0:
+            outputs.append(v)
+    graph.output(outputs[0] if len(outputs) == 1 else tuple(outputs))
+    
+    
+    if print_graph:
+        print(graph)
+    print("starting creating module")
+    gm = GraphModule({}, graph)
+    graph = legalize_graph(gm)
+    gm.graph.lint()
+    print("starting drawing")
+    draw_graph(gm, fname, clear_meta=False)
+
+graph_dump_index = 0
+
 class Scheduler:
     def __init__(self, nodes):
         super(Scheduler, self).__init__()
@@ -448,6 +550,11 @@ class Scheduler:
             else:
                 assert False, node
         self.name_to_node = {node.get_name(): node for node in self.nodes}
+
+        if bool(os.environ.get('INDUCTOR_DEBUG', False)):
+            global graph_dump_index
+            create_fx_graph(self.nodes, f"compute_buffer_{graph_dump_index}", print_graph=True)
+            graph_dump_index += 1
 
         # some new constants could have been created above
         self.available_buffer_names.update(V.graph.constants.keys())
